+++ src/odoc/extract_code.mli
@@ -1,0 +1,7 @@
+val extract :
+  dst:string option ->
+  input:string ->
+  names:string list ->
+  line_directives:bool ->
+  warnings_options:Odoc_model.Error.warnings_options ->
+  (unit, [> `Msg of string ]) result